+++ src/lib.rs
@@ -7,18 +7,13 @@
 //! other metadata fields (album, genre...).
 //! Analyzing a song is as simple as running `Song::from_path("/path/to/song")`.
 //!
-//! The [analysis](Song::analysis) field of each song is an array of f32, which makes the
-//! comparison between songs easy, by just using euclidean distance (see
-//! [distance](Song::distance) for instance).
+//! The [analysis](Song::analysis) field of each song is an array of f32, which
+//! makes the comparison between songs easy, by just using e.g. euclidean
+//! distance (see [distance](Song::distance) for instance).
 //!
 //! Once several songs have been analyzed, making a playlist from one Song
 //! is as easy as computing distances between that song and the rest, and ordering
 //! the songs by distance, ascending.
-//!
-//! It is also convenient to make plug-ins for existing audio players.
-//! It should be as easy as implementing the necessary traits for [Library].
-//! A reference implementation for the MPD player is available
-//! [here](https://github.com/Polochon-street/blissify-rs)
 //!
 //! # Examples
 //!
@@ -65,9 +60,8 @@
 #![warn(missing_docs)]
 #![warn(rustdoc::missing_doc_code_examples)]
 mod chroma;
-pub mod distance;
-pub mod library;
 mod misc;
+pub mod playlist;
 mod song;
 mod temporal;
 mod timbral;
@@ -78,9 +72,11 @@ extern crate num_cpus;
 #[cfg(feature = "serde")]
 #[macro_use]
 extern crate serde;
+use log::info;
+use std::sync::mpsc;
+use std::thread;
 use thiserror::Error;
 
-pub use library::Library;
 pub use song::{Analysis, AnalysisIndex, Song, NUMBER_FEATURES};
 
 const CHANNELS: u16 = 1;
@@ -94,54 +90,73 @@ pub const FEATURES_VERSION: u16 = 1;
 /// Umbrella type for bliss error types
 pub enum BlissError {
     #[error("error happened while decoding file – {0}")]
-    /// An error happened while decoding an (audio) file
+    /// An error happened while decoding an (audio) file.
     DecodingError(String),
     #[error("error happened while analyzing file – {0}")]
-    /// An error happened during the analysis of the samples by bliss
+    /// An error happened during the analysis of the song's samples by bliss.
     AnalysisError(String),
     #[error("error happened with the music library provider - {0}")]
     /// An error happened with the music library provider.
-    /// Useful to report errors when you implement the [Library] trait.
+    /// Useful to report errors when you implement bliss for an audio player.
     ProviderError(String),
 }
 
 /// bliss error type
 pub type BlissResult<T> = Result<T, BlissError>;
 
-/// Simple function to bulk analyze a set of songs represented by their
-/// absolute paths.
+/// Analyze songs in `paths`, and return the analyzed [Song] objects through an
+/// [mpsc::IntoIter]
 ///
-/// When making an extension for an audio player, prefer
-/// implementing the `Library` trait.
-#[doc(hidden)]
-pub fn bulk_analyze(paths: Vec<String>) -> Vec<BlissResult<Song>> {
-    let mut songs = Vec::with_capacity(paths.len());
+/// Returns an iterator, whose items are a tuple made of
+/// the song path (to display to the user in case the analysis failed),
+/// and a Result<Song>.
+///
+/// * Example:
+/// ```no_run
+/// use bliss_audio::{analyze_paths, BlissResult};
+///
+/// fn main() -> BlissResult<()> {
+///     let paths = vec![String::from("/path/to/song1"), String::from("/path/to/song2")];
+///     for (path, result) in analyze_paths(paths) {
+///         match result {
+///             Ok(song) => println!("Do something with analyzed song {} with title {:?}", song.path.display(), song.title),
+///             Err(e) => println!("Song at {} could not be analyzed. Failed with: {}", path, e),
+///         }
+///     }
+///     Ok(())
+/// }
+/// ```
+pub fn analyze_paths(paths: Vec<String>) -> mpsc::IntoIter<(String, BlissResult<Song>)> {
     let num_cpus = num_cpus::get();
 
-    crossbeam::scope(|s| {
-        let mut handles = Vec::with_capacity(paths.len() / num_cpus);
-        let mut chunk_number = paths.len() / num_cpus;
-        if chunk_number == 0 {
-            chunk_number = paths.len();
-        }
-        for chunk in paths.chunks(chunk_number) {
-            handles.push(s.spawn(move |_| {
-                let mut result = Vec::with_capacity(chunk.len());
-                for path in chunk {
-                    let song = Song::from_path(&path);
-                    result.push(song);
-                }
-                result
-            }));
-        }
+    #[allow(clippy::type_complexity)]
+    let (tx, rx): (
+        mpsc::Sender<(String, BlissResult<Song>)>,
+        mpsc::Receiver<(String, BlissResult<Song>)>,
+    ) = mpsc::channel();
+    if paths.is_empty() {
+        return rx.into_iter();
+    }
+    let mut handles = Vec::new();
+    let mut chunk_length = paths.len() / num_cpus;
+    if chunk_length == 0 {
+        chunk_length = paths.len();
+    }
 
-        for handle in handles {
-            songs.extend(handle.join().unwrap());
-        }
-    })
-    .unwrap();
+    for chunk in paths.chunks(chunk_length) {
+        let tx_thread = tx.clone();
+        let owned_chunk = chunk.to_owned();
+        let child = thread::spawn(move || {
+            for path in owned_chunk {
+                info!("Analyzing file '{}'", path);
+                let song = Song::from_path(&path);
+                tx_thread.send((path.to_string(), song)).unwrap();
+            }
+        });
+        handles.push(child);
+    }
 
-    songs
+    rx.into_iter()
 }
 
 #[cfg(test)]
@@ -161,52 +176,28 @@ mod tests {
     }
 
     #[test]
-    fn test_bulk_analyze() {
-        let results = bulk_analyze(vec![
-            String::from("data/s16_mono_22_5kHz.flac"),
-            String::from("data/s16_mono_22_5kHz.flac"),
-            String::from("nonexistent"),
-            String::from("data/s16_stereo_22_5kHz.flac"),
-            String::from("nonexistent"),
-            String::from("nonexistent"),
-            String::from("nonexistent"),
-            String::from("nonexistent"),
-            String::from("nonexistent"),
-            String::from("nonexistent"),
-            String::from("nonexistent"),
-        ]);
-        let mut errored_songs: Vec<String> = results
-            .iter()
-            .filter_map(|x| x.as_ref().err().map(|x| x.to_string()))
-            .collect();
-        errored_songs.sort_by(|a, b| a.cmp(b));
-
-        let mut analyzed_songs: Vec<String> = results
-            .iter()
-            .filter_map(|x| {
-                x.as_ref()
-                    .ok()
-                    .map(|x| x.path.to_str().unwrap().to_string())
+    fn test_analyze_paths() {
+        let paths = vec![
+            String::from("./data/s16_mono_22_5kHz.flac"),
+            String::from("./data/white_noise.flac"),
+            String::from("definitely-not-existing.foo"),
+            String::from("not-existing.foo"),
+        ];
+        let mut results = analyze_paths(paths)
+            .map(|x| match &x.1 {
+                Ok(s) => (true, s.path.to_string_lossy().to_string()),
+                Err(_) => (false, x.0.to_owned()),
             })
-            .collect();
-        analyzed_songs.sort_by(|a, b| a.cmp(b));
-
+            .collect::<Vec<_>>();
+        results.sort();
         assert_eq!(
+            results,
             vec![
-                String::from(
-                    "error happened while decoding file – while opening format: ffmpeg::Error(2: No such file or directory)."
-                );
-                8
+                (false, String::from("definitely-not-existing.foo")),
+                (false, String::from("not-existing.foo")),
+                (true, String::from("./data/s16_mono_22_5kHz.flac")),
+                (true, String::from("./data/white_noise.flac")),
             ],
-            errored_songs
-        );
-        assert_eq!(
-            vec![
-                String::from("data/s16_mono_22_5kHz.flac"),
-                String::from("data/s16_mono_22_5kHz.flac"),
-                String::from("data/s16_stereo_22_5kHz.flac"),
-            ],
-            analyzed_songs,
         );
     }
 }